chore: update eslint config and remove fake Roc
BREAKING CHANGE: remove fake roc api
