--- src/base/BaseRocQuery.ts
@@ -1,26 +1,0 @@
-import { IQueryOptions, IQueryResult, PromisedQueryResult } from '../types';
-
-export default abstract class BaseRocQuery<
-  KeyType = any,
-  ValueType = any,
-  ContentType = Record<string, any>
-> {
-  public readonly viewName: string;
-  protected baseOptions: IQueryOptions;
-  public constructor(viewName: string, options: IQueryOptions) {
-    this.viewName = viewName;
-    this.baseOptions = options;
-  }
-
-  public then(
-    resolve: (
-      value: Array<IQueryResult<KeyType, ValueType, ContentType>>,
-    ) => void,
-    reject: (error: Error) => void,
-  ) {
-    this.fetch().then(resolve, reject);
-  }
-  public abstract fetch(
-    options?: IQueryOptions,
-  ): PromisedQueryResult<KeyType, ValueType, ContentType>;
-}
