--- src/base/index.ts
@@ -1,6 +1,0 @@
-import BaseRoc from './BaseRoc';
-import BaseRocDocument from './BaseRocDocument';
-import BaseRocQuery from './BaseRocQuery';
-import BaseRocReduceQuery from './BaseRocReduceQuery';
-
-export { BaseRoc, BaseRocDocument, BaseRocQuery, BaseRocReduceQuery };
