+++ src/index.ts
@@ -1,4 +1,2 @@
-export * from './fake';
-export * from './real';
-export * from './base';
+export * from './roc';
 export * from './types';
